Refinements to be merged but the API will change drastically (#1)
* Rowx concept and implementation - dropped!!!

Implementing Rowx base type seems to make code more complicated and
invasive for the simple types mapping tables. This makes them harder to
regenerate in the future, when database migration is implemented.

0. Stay in the Modelx implementation. It seems to work already well to
   manage a slice of instances of any struct containing public fields.
   It also keeps close to sqlx simplicity of usage and just wraps
   preparatory and otherwise repetitive code around it. Stay simple for
   the user. Do not distract him from the business logic as much as
   possible.

1. Implemented Get for getting single records from the database.

Next TODOS?:
1. Implement PrimaryKeys method?... To be used in Insert to not insert
   the default values from go, but allow the database to autoincrement
   or run it's triggers, etc.
2. Implement UniqueKeys method?..

* Improved documentation. Update panics now if no data is provided.

* Updated and tidied dependencies

* Solved problem with updating!!! WIP.

Solution: Use a public field with a substructure as binding parameters.
Do not use this field as a column - that's it.
Next step: Finish the concept with using field tags and move to the
controllers module for slovo.

* WIP reduced methods - working state.

* cleanup

* fix comment

* Now the WHERE clause must not be written in the where parameter.

* WHERE is just optional now.

* Added dummy benchmarks and fuzz tests...

Adding Examples - WIP

* Adding Examples - WIP

* Examples WIP
TODO:
Change the package name to rx
Check at runtime if the SqlxRows structures implement methods Table() and
Columns(). Use them in Modelx if the generic type imlenets the
respective method.
Next: implement the TODO