Renamed a lot of functions, optimized code, cleaned up documentation. (#2)
* Added more linters for better code and self discipline.
* Replaced remaining occurrences of Modelx with Rx.

Fixed some remaining linters' errors.
Added function SnakeToCamel. Will be used to convert table and column names to
structure and field names.
Moved Interfaces to rx/meta.go. Added new interface SqlxMeta.
Renamed SqlxRows to Rowx to remind me why I named the library Rowx - extended
structures with methods which know that they are rows in a database and have
Table() and Columns(). But it is not needed now.
* Renamed utility functions and optimised SnakeToCamel
* Small changes in the documentation.
* Added README